refactor: replace getMedicationDisplayName with getHumanReadableCodeableConcept update related tests, and move
