+++ containers/ecr-viewer/src/app/services/formatService.ts
@@ -1,6 +1,6 @@
 import "server-only"; // FHIR evaluation/formatting should be done server side
 
-import { Address, ContactPoint, HumanName } from "fhir/r4";
+import { Address, CodeableConcept, ContactPoint, HumanName } from "fhir/r4";
 import { toSentenceCase, toTitleCase } from "@/app/utils/format-utils";
 import { formatDate } from "./formatDateService";
 
@@ -208,4 +208,47 @@ export const formatContactPoint = (
     }
   }
   return contactArr.join("\n");
+};
+
+/**
+ * Attempts to return a human-readable display value for a CodeableConcept. It will return the first
+ * available value in the following order:
+ * 1) `undefined` if the `CodeableConcept` is falsy
+ * 2) `CodeableConcept.text`
+ * 3) value of the first `coding` with a `display` value
+ * 4) `code` and `system` values of the first `coding` with a `code` and `system values.
+ * 5) `code` of the first `coding` with a `code` value
+ * 6) `undefined`
+ * @param codeableConcept - The CodeableConcept to get the display value from.
+ * @returns - The human-readable display value of the CodeableConcept.
+ */
+export const getHumanReadableCodeableConcept = (
+  codeableConcept: CodeableConcept | undefined,
+) => {
+  if (!codeableConcept) {
+    return undefined;
+  }
+
+  const { coding, text } = codeableConcept;
+
+  if (text) {
+    return text;
+  }
+
+  const firstCodingWithDisplay = coding?.find((c) => c.display);
+  if (firstCodingWithDisplay?.display) {
+    return firstCodingWithDisplay.display;
+  }
+
+  const firstCodingWithCodeSystem = coding?.find((c) => c.code && c.system);
+  if (firstCodingWithCodeSystem?.code && firstCodingWithCodeSystem?.system) {
+    return `${firstCodingWithCodeSystem.code} (${firstCodingWithCodeSystem.system})`;
+  }
+
+  const firstCodingWithCode = coding?.find((c) => c.code);
+  if (firstCodingWithCode?.code) {
+    return firstCodingWithCode.code;
+  }
+
+  return undefined;
 };
