+++ containers/ecr-viewer/src/app/tests/assets/BundleEcrMetadata.json
@@ -499,7 +499,8 @@
                 "coding": [
                   {
                     "system": "http://terminology.hl7.org/CodeSystem/v3-ParticipationType",
-                    "code": "ATND"
+                    "code": "ATND",
+                    "display": "attender"
                   }
                 ]
               }
